+++ source/ChanSort/GlobalImageCollection.cs
@@ -17,7 +17,7 @@ namespace ChanSort.Ui
   /// that are used globally throughout the solution
   /// </summary>
   [ToolboxItem(false)]
-  public class GlobalImageCollectionHolder : XtraForm //Component - DX23.1 it must be an XtraForm because the context menu for editing the collections is missing in the Designer
+  public class GlobalImageCollectionHolder : Component
   {
     private IContainer components;
     private ImageCollection rawImageCollection;
@@ -47,7 +47,6 @@ namespace ChanSort.Ui
       System.ComponentModel.ComponentResourceManager resources = new System.ComponentModel.ComponentResourceManager(typeof(GlobalImageCollectionHolder));
       this.rawImageCollection = new DevExpress.Utils.ImageCollection(this.components);
       ((System.ComponentModel.ISupportInitialize)(this.rawImageCollection)).BeginInit();
-      this.SuspendLayout();
       // 
       // rawImageCollection
       // 
@@ -104,13 +103,7 @@ namespace ChanSort.Ui
       this.rawImageCollection.Images.SetKeyName(49, "0049.png");
       this.rawImageCollection.Images.SetKeyName(50, "0050.png");
       this.rawImageCollection.Images.SetKeyName(51, "0051.png");
-      // 
-      // GlobalImageCollectionHolder
-      // 
-      this.ClientSize = new System.Drawing.Size(298, 268);
-      this.Name = "GlobalImageCollectionHolder";
       ((System.ComponentModel.ISupportInitialize)(this.rawImageCollection)).EndInit();
-      this.ResumeLayout(false);
 
     }