+++ source/Information/FileStructures_for_HHD_Hex_Editor_Neo/dtv-cmdb.h
@@ -94,7 +94,7 @@ struct s_Channel
 };
 
 
-public struct dtv_cmdb_2
+public struct dtv_cmdb_2_unified
 {
   var headerLength = 0;
   var channelBitmapLength = 0;